Refactor property handling and introduce ToolMainProperty
Replaced ToolNonPrimitiveProperty with ToolMainProperty in SceneBuilder.cs and ActorsOpenAiEndpointReader.cs. Modified ToolNonPrimitiveProperty.cs to be non-sealed and removed properties/methods related to Required and AdditionalProperties. Added new ToolMainProperty.cs file defining ToolMainProperty class with Required and AdditionalProperties management.
