+++ src/PlayFramework/Rystem.PlayFramework/Chat/Models/Tools/ToolNonPrimitiveProperty.cs
@@ -2,7 +2,7 @@
 
 namespace Rystem.PlayFramework
 {
-    public sealed class ToolNonPrimitiveProperty : ToolProperty
+    public class ToolNonPrimitiveProperty : ToolProperty
     {
         internal const string DefaultTypeName = "object";
         public ToolNonPrimitiveProperty()
@@ -12,16 +12,6 @@ namespace Rystem.PlayFramework
         }
         [JsonPropertyName("properties")]
         public Dictionary<string, ToolProperty> Properties { get; }
-        [JsonPropertyName("required")]
-        public List<string>? Required { get; private set; }
-        [JsonPropertyName("additionalProperties")]
-        public bool AdditionalProperties { get; set; }
-        public ToolNonPrimitiveProperty AddRequired(params string[] names)
-        {
-            Required ??= new List<string>();
-            Required.AddRange(names);
-            return this;
-        }
         public ToolNonPrimitiveProperty AddEnum(string key, ToolEnumProperty property)
             => AddProperty(key, property);
         public ToolNonPrimitiveProperty AddObject(string key, ToolNonPrimitiveProperty property)
